pyproject.toml
[project]
name = "nl-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.14.0",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asyncpg==0.30.0",
    "bcrypt==4.2.1",
    "black==24.10.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.115.6",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "mako==1.3.7",
    "markupsafe==3.0.2",
    "mypy-extensions==1.0.0",
    "packaging==24.2",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "psycopg2==2.9.10",
    "pydantic-core==2.27.1",
    "pydantic==2.10.3",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.19",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "starlette==0.41.3",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.1",
    "langchain-community>=0.3.12",
    "langchain-openai>=0.2.12",
    "nemoguardrails>=0.11.0",
    "redis>=5.2.1",
    "aiosmtplib>=3.0.2",
]
